+++ Sora/PeerChannel.swift
@@ -581,7 +581,8 @@ class BasicPeerChannelContext: NSObject, RTCPeerConnectionDelegate {
             audioEnabled: true,
             audioCodec: configuration.audioCodec,
             audioBitRate: configuration.audioBitRate,
-            spotlight: configuration.spotlight,
+            spotlightEnabled: configuration.spotlightEnabled,
+            activeSpeakerLimit: configuration.activeSpeakerLimit,
             simulcastEnabled: configuration.simulcastEnabled,
             simulcastQuality: configuration.simulcastQuality,
             soraClient: soraClient,
@@ -633,6 +634,8 @@ class BasicPeerChannelContext: NSObject, RTCPeerConnectionDelegate {
                 
                 // カテゴリをマイク用途のものに変更する
                 // libwebrtc の内部で参照される RTCAudioSessionConfiguration を使う必要がある
+                Logger.debug(type: .peerChannel,
+                             message: "change audio session category (playAndRecord)")
                 RTCAudioSessionConfiguration.webRTC().category =
                     AVAudioSession.Category.playAndRecord.rawValue
                 
@@ -644,7 +647,7 @@ class BasicPeerChannelContext: NSObject, RTCPeerConnectionDelegate {
                     }
                     self.isAudioInputInitialized = true
                     Logger.debug(type: .peerChannel,
-                                 message: "audio input is initialized")
+                                 message: "audio input is initialized => category \(RTCAudioSession.sharedInstance().category)")
                 }
             }
         }